feat: Add quest ChaseARainbowQuest and specify claiming logic (#53)
Adds the Chase a Rainbow Quest, giving the player Prismatic Shard. When they collect cards from 3 unique colors, they can claim the quest to lose Prismatic Shard and add it and Question Card to a reward screen (so the player may choose to keep Prismatic Shard, or leave it). Will change to make Prismatic Shard Questbound when it is added. When 2+ character mods are installed, the number will dynamically increase.
Also adds rewardScreenOnly boolean for quests that can only be claimed while there is a reward screen

* Setup

Add author name to list of authors and create new files copied from example files.

* rewardScreenOnly added

Added rewardsScreenOnly variable that makes it so that quests that have it set to true can only be claimed if there is a reward screen already. (Set to false by default)

* Rainbow Quest!

Fully added the rainbow quest. For now always requires 3 cards, might change to be dynamic based on amount of colors if possible.

* Can't Claim Message

Your character will now complain and explain why when you try to claim a quest that you can not claim.

* Completed Text Claim timing

The Quest Completed texts now mention when you can claim them, outside of combat for normal quests and at a reward screen for quests that can only be claimed then.

* Dynamic Req

Updates Rainbow Quest to dynamically determine the number of required different colors based on the amount of found colors. 3 by base, increases to 4 at 7 colors (so 2 characters added).

* CanSpawn

Forgot to add the CanSpawn condition. Now should not be able to spawn if the player has the Multiquest Quest or Emblem, and not when they already have a Prismatic Shard.

* Some feedback implemented

---------

Co-authored-by: erasels <emanuelbedburdick@yahoo.de>